Extract interface for FunctionDependencies
